fixes after Marco's review
